examples/04_local_slm_kv_extraction/requirements.txt
llama-cpp-python>=0.3.0
huggingface-hub>=0.23.0
